onequant/requirements.txt
websockets==14.2
aiohttp==3.11.11
python-dotenv==1.0.1
